pyproject.toml
[project]
name = "pdf-to-tables-upwork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marker-pdf>=1.5.4",
]
